# apps/api/openapi/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/stats": {
      "get": {
        "description": "Returns live counts, phase mix, weekly handoffs, sub-team load, pipeline steps, and recent files for the signed-in person's teams.",
        "operationId": "getStats",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Dashboard stats.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DashboardStatsDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get floor stats",
        "tags": [
          "Projects",
          "Stats"
        ]
      }
    },
    "/projects": {
      "get": {
        "description": "Admins and team leads see active projects on their teams. Members see projects currently in their sub-team.",
        "operationId": "getProjects",
        "parameters": [
          {
            "name": "teamId",
            "required": false,
            "in": "query",
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": "Projects in progress."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List work",
        "tags": [
          "Projects"
        ]
      },
      "post": {
        "description": "Team leads create a project. Plugin issues a project ID and opens the first phase for its sub-team.",
        "operationId": "postProjects",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProjectDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Project created."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a project",
        "tags": [
          "Projects"
        ]
      }
    },
    "/projects/{id}": {
      "get": {
        "description": "Returns the project, every phase ID, and uploaded files.",
        "operationId": "getProjectsById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Project id",
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": "Project detail."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get a project",
        "tags": [
          "Projects"
        ]
      }
    },
    "/project-phases/{id}/documents": {
      "post": {
        "description": "Adds one or more files to the active sub-team phase. Files are stored in R2 (or local disk in development).",
        "operationId": "postProjectPhaseDocuments",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Project phase id",
            "schema": {}
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "files": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "format": "binary"
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Files saved."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Upload files to a phase",
        "tags": [
          "Projects"
        ]
      }
    },
    "/project-phases/{id}/complete": {
      "post": {
        "description": "Marks the sub-team's work complete. If every parallel phase on this step is done, the next step starts.",
        "operationId": "postProjectPhaseComplete",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Project phase id",
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": "Phase handed off."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Hand off a phase",
        "tags": [
          "Projects"
        ]
      }
    },
    "/projects/{id}/send-back": {
      "post": {
        "description": "Reopens the previous step with a new phase ID and keeps previous files as history.",
        "operationId": "postProjectsSendBack",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Project id",
            "schema": {}
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SendBackDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Work sent back."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Send work backwards",
        "tags": [
          "Projects"
        ]
      }
    },
    "/history": {
      "get": {
        "description": "Returns phases a sub-team has worked on, with their IDs and files.",
        "operationId": "getHistory",
        "parameters": [],
        "responses": {
          "200": {
            "description": "History."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List sub-team history",
        "tags": [
          "Projects",
          "History"
        ]
      }
    },
    "/documents/{id}/file": {
      "get": {
        "description": "Streams a previously uploaded handoff document.",
        "operationId": "getDocumentsFile",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Document id",
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Download a file",
        "tags": [
          "Projects"
        ]
      }
    }
  },
  "info": {
    "title": "Plugin Project Manager API",
    "description": "API for Plugin teams to create projects, move work through configurable phases, and hand off documents.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "/api/v1"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "DashboardOldestJobDto": {
        "type": "object",
        "properties": {
          "projectId": {
            "type": "string"
          },
          "publicId": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "phaseName": {
            "type": "string"
          },
          "subTeamName": {
            "type": "string"
          },
          "daysWaiting": {
            "type": "number"
          }
        },
        "required": [
          "projectId",
          "publicId",
          "name",
          "phaseName",
          "subTeamName",
          "daysWaiting"
        ]
      },
      "DashboardWaitingItemDto": {
        "type": "object",
        "properties": {
          "projectId": {
            "type": "string"
          },
          "projectPublicId": {
            "type": "string"
          },
          "projectName": {
            "type": "string"
          },
          "phasePublicId": {
            "type": "string"
          },
          "phaseName": {
            "type": "string"
          },
          "subTeamName": {
            "type": "string"
          },
          "teamName": {
            "type": "string"
          },
          "daysWaiting": {
            "type": "number"
          }
        },
        "required": [
          "projectId",
          "projectPublicId",
          "projectName",
          "phasePublicId",
          "phaseName",
          "subTeamName",
          "teamName",
          "daysWaiting"
        ]
      },
      "DashboardSubTeamLoadDto": {
        "type": "object",
        "properties": {
          "subTeamId": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "teamName": {
            "type": "string"
          },
          "activeCount": {
            "type": "number"
          }
        },
        "required": [
          "subTeamId",
          "name",
          "teamName",
          "activeCount"
        ]
      },
      "DashboardPipelineStepDto": {
        "type": "object",
        "properties": {
          "step": {
            "type": "number"
          },
          "label": {
            "type": "string"
          },
          "projectCount": {
            "type": "number"
          }
        },
        "required": [
          "step",
          "label",
          "projectCount"
        ]
      },
      "DashboardHandoffDto": {
        "type": "object",
        "properties": {
          "projectId": {
            "type": "string"
          },
          "projectPublicId": {
            "type": "string"
          },
          "projectName": {
            "type": "string"
          },
          "phasePublicId": {
            "type": "string"
          },
          "phaseName": {
            "type": "string"
          },
          "subTeamName": {
            "type": "string"
          },
          "completedAt": {
            "type": "string"
          }
        },
        "required": [
          "projectId",
          "projectPublicId",
          "projectName",
          "phasePublicId",
          "phaseName",
          "subTeamName",
          "completedAt"
        ]
      },
      "DashboardPhaseMixDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "example": "active"
          },
          "count": {
            "type": "number"
          }
        },
        "required": [
          "status",
          "count"
        ]
      },
      "DashboardDayCountDto": {
        "type": "object",
        "properties": {
          "day": {
            "type": "string",
            "example": "2026-09-10"
          },
          "label": {
            "type": "string",
            "example": "Wed"
          },
          "count": {
            "type": "number"
          }
        },
        "required": [
          "day",
          "label",
          "count"
        ]
      },
      "DashboardStatsDto": {
        "type": "object",
        "properties": {
          "activeProjects": {
            "type": "number"
          },
          "completedProjects": {
            "type": "number"
          },
          "completedThisWeek": {
            "type": "number"
          },
          "waitingOnYou": {
            "type": "number"
          },
          "returnedPhases": {
            "type": "number"
          },
          "filesUploaded": {
            "type": "number"
          },
          "oldestJob": {
            "nullable": true,
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/DashboardOldestJobDto"
              }
            ]
          },
          "waiting": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DashboardWaitingItemDto"
            }
          },
          "bySubTeam": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DashboardSubTeamLoadDto"
            }
          },
          "pipeline": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DashboardPipelineStepDto"
            }
          },
          "recentHandoffs": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DashboardHandoffDto"
            }
          },
          "phaseMix": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DashboardPhaseMixDto"
            }
          },
          "handoffsByDay": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DashboardDayCountDto"
            }
          }
        },
        "required": [
          "activeProjects",
          "completedProjects",
          "completedThisWeek",
          "waitingOnYou",
          "returnedPhases",
          "filesUploaded",
          "oldestJob",
          "waiting",
          "bySubTeam",
          "pipeline",
          "recentHandoffs",
          "phaseMix",
          "handoffsByDay"
        ]
      },
      "CreateProjectDto": {
        "type": "object",
        "properties": {
          "teamId": {
            "type": "string"
          },
          "name": {
            "type": "string",
            "example": "Nestle Q4 activation"
          },
          "description": {
            "type": "string"
          }
        },
        "required": [
          "teamId",
          "name"
        ]
      },
      "SendBackDto": {
        "type": "object",
        "properties": {
          "targetPhaseId": {
            "type": "string",
            "description": "Project phase id to reopen."
          },
          "reason": {
            "type": "string"
          }
        },
        "required": [
          "targetPhaseId"
        ]
      }
    }
  }
}